requirements.txt
# Pacotes principais
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.3.0
prophet==1.1.4
yfinance==0.2.28

# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
matplotlib==3.7.2
plotly==5.16.0